Update YesSQL to 3.2.0 (#12528)

+++ src/OrchardCore/OrchardCore.Autoroute.Core/Services/AutorouteEntries.cs
@@ -22,7 +22,7 @@ namespace OrchardCore.Autoroute.Core.Services
         private ImmutableDictionary<string, AutorouteEntry> _contentItemIds = ImmutableDictionary<string, AutorouteEntry>.Empty;
         private readonly SemaphoreSlim _semaphore = new SemaphoreSlim(1);
 
-        private int _lastIndexId;
+        private long _lastIndexId;
         private string _stateIdentifier;
         private bool _initialized;
